Split up the item module into smaller modules
- Created bodyElement module
- Created liveBlock module
- Created docParser module to house the DocParser type
- Migrated the CAPI date parser to the capi module
- Updated imports
