+++ src/types/docParser.ts
@@ -1,0 +1,8 @@
+// ----- Type ----- //
+
+type DocParser = (html: string) => DocumentFragment;
+
+
+// ----- Exports ----- //
+
+export default DocParser;
